Guard quantum armor client-only logic
